feat(HomeBrew): AND-1350 Live quotes - move quote service
