+++ morph/common/src/main/java/com/blockchain/morph/exchange/service/QuoteService.kt
@@ -1,4 +1,4 @@
-package com.blockchain.morph.homebrew
+package com.blockchain.morph.exchange.service
 
 import com.blockchain.morph.exchange.mvi.Quote
 import com.blockchain.morph.quote.ExchangeQuoteRequest
